Add: Show, Buttons

+++ src/main/java/it/lessons/pizzeria/repository/PizzaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import it.lessons.pizzeria.model.Pizza;
 
-public interface PizzaRepository extends JpaRepository<Pizza, Integer>{
+public interface PizzaRepository extends JpaRepository<Pizza, Long>{
 	
 }
